Update integration tests as test target
